* new: add interaction + new room. fix: gui + click + interaction

+++ Assets/Script/Battle/Item/Arming/Ammunition.cs
@@ -3,14 +3,14 @@ using System.Collections;
 
 public abstract class Ammunition
 {
-    protected int damage;
+    protected float damage;
     protected int weight;
 
     protected Ammunition()
     {
     }
 
-    public int getDamage()
+    public float getDamage()
     {
         return this.damage;
     }
